+++ collect-server/src/main/java/org/openforis/collect/metamodel/proxy/TaxonSummaryProxy.java
@@ -67,7 +67,7 @@ public class TaxonSummaryProxy implements Proxy {
 	}
 
 	@ExternalizedProperty
-	public int getTaxonSystemId() {
+	public long getTaxonSystemId() {
 		return summary.getTaxonSystemId();
 	}
